+++ app/domain/authentication/audit_event/login.rb
@@ -1,15 +1,15 @@
 module Authentication
   module AuditEvent
-    class Authenticate < ::Audit::Event::Authn   
-      operation 'authenticate'
-    
+    class Login < ::Audit::Event::Authn  
+      operation 'login'
+
       success_message do
-        "#{role_id} successfully authenticated with authenticator "\
+        "#{role_id} successfully logged in with authenticator "\
           "#{authenticator_name}#{service_message_part}"
       end
-    
+
       failure_message do
-        "#{role_id} failed to authenticate with authenticator "\
+        "#{role_id} failed to login with authenticator "\
           "#{authenticator_name}#{service_message_part}"
       end
     end