+++ app/domain/authentication/audit_event/validate_status.rb
@@ -1,15 +1,15 @@
 module Authentication
   module AuditEvent
-    class Authenticate < ::Audit::Event::Authn   
-      operation 'authenticate'
-    
+    class ValidateStatus < ::Audit::Event::Authn  
+      operation 'validate-status'
+
       success_message do
-        "#{role_id} successfully authenticated with authenticator "\
+        "#{role_id} successfully validated status for authenticator "\
           "#{authenticator_name}#{service_message_part}"
       end
-    
+
       failure_message do
-        "#{role_id} failed to authenticate with authenticator "\
+        "#{role_id} failed to validate status for authenticator "\
           "#{authenticator_name}#{service_message_part}"
       end
     end