+++ app/domain/authentication/validate_status.rb
@@ -83,7 +83,7 @@ module Authentication
 
     def audit_success
       @log_audit_event.(
-        event: ::Authentication::AuditEvent::Authenticate,
+        event: ::Authentication::AuditEvent::ValidateStatus,
         authenticator_input: @authenticator_status_input,
         success: true,
         message: nil
@@ -92,7 +92,7 @@ module Authentication
 
     def audit_failure(err)
       @log_audit_event.(
-        event: ::Authentication::AuditEvent::Authenticate,
+        event: ::Authentication::AuditEvent::ValidateStatus,
         authenticator_input: @authenticator_status_input,
         success: false,
         message: err.message